dtype

+++ sklearn/cluster/_bisect_k_means.py
@@ -643,8 +643,8 @@ class BisectKMeans(KMeans):
 
         # sample_weight is needed for labeling in _check_labels_threadpool_limit()
         # as argument, but not used there since its not updating centers.
-        # In this case, it doesn't have to be shape of X - only must be ndarray
-        sample_weight = np.zeros(1)
+        # In this case, it doesn't have to be shape of X - only ndarray wits same dtype
+        sample_weight = np.zeros(1, dtype=X.dtype)
 
         # With only one cluster all points have same label
         if self.n_clusters == 1:
@@ -670,6 +670,7 @@ class BisectKMeans(KMeans):
             for idx in centers_hierarchical:
                 center_ids = [idx]
 
+                # If center has children - relabel its data to child centers
                 if self._inner_tree[idx]["children"] is not None:
                     center_ids = self._inner_tree[idx]["children"]
                     picked_samples = labels == idx
